+++ CHANGELOG.md
@@ -10,9 +10,10 @@
 >   more of my free time to fixing bugs and developing new features 🤗
 
 ## v0.9.4 (not released)
-- **bugfix**: [Error when using hooks defined as a subclass of LiveView's v1.1 new ViewHook](https://github.com/phenixdigital/phoenix_storybook/issues/663)
 - **feature**: [support custom light class in color mode sandbox](https://github.com/phenixdigital/phoenix_storybook/issues/722)
+- **feature**: allow live_component stories to implement `handle_info/2` for parent message handling (e.g. `send_update/2`)
 - **bugfix**: [storybook asset hash query param renamed to avoid live reload URL rewriting](https://github.com/phenixdigital/phoenix_storybook/issues/699)
+- **bugfix**: [Error when using hooks defined as a subclass of LiveView's v1.1 new ViewHook](https://github.com/phenixdigital/phoenix_storybook/issues/663)
 
 ## v0.9.3 (2025-09-15)
 - **feature**: [accompany variations with optional notes](https://github.com/phenixdigital/phoenix_storybook/issues/571) - 🙏 [@xxdavid](https://github.com/xxdavid)